requirements/base.txt
django>=4.2,<5.0
celery[redis]>=5.3
redis>=5.0
psycopg2-binary>=2.9
django-environ>=0.11
django-extensions>=3.2
django-celery-beat>=2.6
whitenoise>=6.6
yt-dlp>=2024.3
google-cloud-storage>=2.14
google-cloud-aiplatform>=1.50
google-generativeai>=0.5
pinecone>=5.0
flower>=2.0
reportlab>=4.0
